refactor: delegate lightbox prefetching to CloudinaryImage
- Removed `imageQueries.ts` and encapsulated all sizing options, url resolvers, and query functions inside `CloudinaryImage.tsx` to maintain clean component boundaries.
- Replaced the manual image-preload effect in `PieceDetail.tsx` with parallel prefetching triggered via an internal `useEffect` inside `CloudinaryImage`.
- Updated custom `render` wrapper in `CloudinaryImage.test.tsx` to supply the `QueryClientProvider` and wrap `rerender` to retain query context.
- Refactored `ImageLightbox.test.tsx` to use a component-level mock of `SuspenseCloudinaryImage` to cleanly simulate suspended and loaded render states without timing dependencies.
- Deleted dead `imageQueries.ts` files and build dependencies.
