+++ web/src/components/__tests__/PiecePhotoGalleryLightboxIntegration.test.tsx
@@ -23,12 +23,6 @@ vi.mock("../CloudinaryImage", () => ({
   ),
 }));
 
-vi.mock("../../util/imageQueries", () => ({
-  useSuspendedImageLoad: () => ({ data: "mock-url" }),
-  imageLoadQueryOptions: (url: string) => ({ queryKey: ["image-load", url] }),
-  getCloudinaryUrl: ({ url }: { url: string }) => url,
-}));
-
 vi.mock("masonic", () => ({
   Masonry: ({
     items,
